+++ test/foundry/Raffle.claimRefund.t.sol
@@ -109,7 +109,7 @@ contract Raffle_ClaimRefund_Test is TestHelpers {
 
             assertEq(participant.balance, 0.025 ether);
 
-            (uint256 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(
+            (uint256 amountPaid, uint80 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(
                 0,
                 participant
             );
